% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse.r
\name{create_rtm_docs}
\alias{create_rtm_docs}
\title{Convert HTML to documentation}
\usage{
create_rtm_docs(html_pages)
}
\arguments{
\item{html_pages}{a list of html documentation objects}
}
\value{
a named list of rtm_doc objects
}
\description{
Convert HTML to documentation
}

