% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/delay.r
\name{delay_by}
\alias{delay_by}
\title{Ensures that a certain amount of time has elapsed before
a function can be called.}
\usage{
delay_by(delay, f)
}
\arguments{
\item{delay}{the delay in seconds}

\item{f}{the existing function}
}
\value{
the new function, which behaves exactly as the
  existing function except it calls Sys.sleep when necessary
}
\description{
This is a function operator, so you specify the delay
and an existing function and it returns a new function
that is only callable once every \code{delay} seconds.
Suggested in an exercise in \href{http://adv-r.had.co.nz/}{"Advanced R"}
}

