% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetch.r
\name{fetch_html_docs}
\alias{fetch_html_docs}
\title{Get a list of html documentation for all RTM API method calls from
the RTM website.}
\usage{
fetch_html_docs(rtm_url = "https://www.rememberthemilk.com/",
  api_doc_path = "services/api/methods/", css = "#api a", progress = TRUE,
  sample = -1)
}
\arguments{
\item{rtm_url}{the base url of the RTM website}

\item{api_doc_path}{the path to the list of links to documentation on
each method call}

\item{css}{the css selector for the links}

\item{progress}{whether to show a progress bar}

\item{sample}{select a random sample of this size, for testing,
by default -1 means scrape all documentation}
}
\value{
a list of html pages
}
\description{
Get a list of html documentation for all RTM API method calls from
the RTM website.
}

