% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/delay.r
\name{html_d}
\alias{html_d}
\title{An instance of the \code{rvest::html()} function that can be
run at most once every 1 second.}
\usage{
html_d(...)
}
\arguments{
\item{...}{same parameters as \code{rvest::html()}}
}
\value{
same as \code{rvest::html()}
}
\description{
An instance of the \code{rvest::html()} function that can be
run at most once every 1 second.
}

