% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse.r
\name{print.rtm_doc}
\alias{print.rtm_doc}
\title{Print a description.}
\usage{
\method{print}{rtm_doc}(x, ...)
}
\arguments{
\item{x}{the RTM object}

\item{...}{extra}
}
\description{
Print a description.
}

