% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rtmapidoc.r
\docType{package}
\name{rtmapidoc}
\alias{rtmapidoc}
\alias{rtmapidoc-package}
\title{\pkg{rtmapidoc} scrapes the Remember the Milk API documentation website
for the purposes of creating and maintaining the rtmr package}
\description{
\pkg{rtmapidoc} scrapes the Remember the Milk API documentation website
for the purposes of creating and maintaining the rtmr package
}

